+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Compatibility:
 * Implement specializations for immutable ruby objects for ObjectSpace methods (@bjfish).
 * Use `$PAGER` for `--help` and `--help*`, similar to CRuby (#2542, @Strech).
 * Ensure all headers are warnings-free (#2662, @eregon).
+* All `IO` instances should have `T_FILE` as their `rb_type()`, not only `File` instances (#2662, @eregon).
 
 Performance: